ajustes en Usuario para las excepciones personalizadas

+++ src/main/java/cl/td/g2/eventos/service/UsuarioService.java
@@ -1,6 +1,8 @@
 package cl.td.g2.eventos.service;
 
 import cl.td.g2.eventos.dto.UsuarioDTO;
+import cl.td.g2.eventos.exception.BadRequestException;
+import cl.td.g2.eventos.exception.NotFoundException;
 import cl.td.g2.eventos.mapper.UsuarioMapper;
 import cl.td.g2.eventos.model.Usuario;
 import cl.td.g2.eventos.repository.UsuarioRepository;
@@ -8,7 +10,6 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
-import java.util.Optional;
 
 @Service
 public class UsuarioService {
@@ -21,30 +22,50 @@ public class UsuarioService {
 
     public List<UsuarioDTO> getAllUsuarios() {
         List<Usuario> usuarios = usuarioRepository.findAll();
+        if (usuarios == null || usuarios.isEmpty()) {
+        	throw new NotFoundException("Usuarios");
+        }
         return usuarioMapper.toDTO(usuarios);
     }
 
-    public Optional<UsuarioDTO> getUsuarioById(Long id) {
-        return usuarioRepository.findById(id).map(usuarioMapper::toDTO);
+    public UsuarioDTO getUsuarioById(Long id) {
+        return usuarioRepository.findById(id).map(usuarioMapper::toDTO).orElseThrow(() -> new NotFoundException("Usuario", "id", id));
     }
 
     public UsuarioDTO createUsuario(UsuarioDTO usuarioDTO) {
         Usuario usuario = usuarioMapper.toEntity(usuarioDTO);
-        Usuario savedUsuario = usuarioRepository.save(usuario);
-        return usuarioMapper.toDTO(savedUsuario);
+        try {
+        	Usuario savedUsuario = usuarioRepository.save(usuario);
+            return usuarioMapper.toDTO(savedUsuario);
+		} catch (Exception ex) {
+			throw new BadRequestException(ex.getMessage());
+		}
     }
 
     public UsuarioDTO updateUsuario(Long id, UsuarioDTO usuarioDTO) {
         if (usuarioRepository.existsById(id)) {
-            Usuario usuario = usuarioMapper.toEntity(usuarioDTO);
+        	Usuario usuario = usuarioMapper.toEntity(usuarioDTO);
             usuario.setId(id);
-            Usuario updatedUsuario = usuarioRepository.save(usuario);
-            return usuarioMapper.toDTO(updatedUsuario);
+            try {
+            	Usuario updatedUsuario = usuarioRepository.save(usuario);
+                return usuarioMapper.toDTO(updatedUsuario);
+    		} catch (Exception ex) {
+    			throw new BadRequestException(ex.getMessage());
+    		}
         }
-        return null;
+        throw new NotFoundException("Usuario", "id", id);
     }
 
     public void deleteUsuario(Long id) {
-        usuarioRepository.deleteById(id);
+        if (usuarioRepository.existsById(id)) {
+    		try {
+    			usuarioRepository.deleteById(id);
+    		} catch (Exception ex) {
+    			throw new BadRequestException(ex.getMessage());
+    		}
+        }
+    	else {
+    		throw new NotFoundException("Usuario", "id", id);
+    	}
     }
 }
